+++ sql/install.sql
@@ -65,7 +65,7 @@ CREATE SEQUENCE car_ID;
 CREATE TABLE Car
 (
   id INT DEFAULT nextval('car_ID') NOT NULL,
-  customer_id LONG NOT NULL,
+  customer_id INT NOT NULL,
   make_year VARCHAR(5) NOT NULL,
   make VARCHAR(60) NOT NULL,
   model VARCHAR(60) NOT NULL,
